refactored user settings page
