+++ frontend/services/entities/version.ts
@@ -1,0 +1,11 @@
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+import sendRequest from "services";
+import endpoints from "fleet/endpoints";
+
+export default {
+  load: () => {
+    const { VERSION } = endpoints;
+    
+    return sendRequest("GET", VERSION);
+  },
+};
